+++ src/richie/apps/courses/urls.py
@@ -1,11 +1,12 @@
 """
 API routes exposed by our Courses app.
 """
-from django.urls import re_path
+from django.urls import path, re_path
 
 from rest_framework import routers
 
 from .api import CourseRunsViewSet, course_runs_sync
+from .views import PersonPageAdminAutocomplete
 
 ROUTER = routers.SimpleRouter()
 
@@ -14,4 +15,9 @@ ROUTER.register("course-runs", CourseRunsViewSet, "course_runs")
 
 urlpatterns = ROUTER.urls + [
     re_path("course-runs-sync/?$", course_runs_sync, name="course_run_sync"),
+    path(
+        "person-page-admin-autocomplete/",
+        PersonPageAdminAutocomplete.as_view(),
+        name="person-page-admin-autocomplete",
+    ),
 ]