+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add course runs to the course search API
+- Add autocomplete feature on Page select in Person plugin
 
 ### Changed